Added Energy Burst and base for Laser Beam
Changed damage variable type from int to float to allow multiplication of damage values. Fixed issue with instantiated projectiles not having the same speed as the weapon creating them. Final alpha version of blaster and energy burst scripts. Updated prefabs and added new charged wave prefab.

+++ Assets/_Game/Scripts/Mechanics/Enemies/EnemyBase.cs
@@ -43,7 +43,7 @@ public abstract class EnemyBase : EntityBase
 
     public abstract void Dead();
 
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(float damage)
     {
 		// Prevent enemies from taking damage multiple times in the same frame
 		if (Time.time - cdInvuln > 0.01f)
